feature: type analysis
* Type Analysis

Add a system for type analysis.
The central element is a list of rules.
Every rule describes a type, that can reference other rules.
The AST is traversed and rules for the different expressions are added.
The underlying rules are not evaluated, just referenced.
This enables cyclic dependencies.
Furthermore assertions about the equality of two rules/types can be made.
The valididy of the assertions is checked afterwards
by comparing the two types.
Asserted are the equality of a type expression
and the initialization expression
as well as the equality between a named type
and its initialization expression.

* Semantic Analysis

Remove queries.
Add type analysis.
Check for multiple inits.

* Other

Add helper functions in `ast`. E.g. `info()` for enums.
Add GlobalIdent in `lib`.
Add error messages for all the new errors.
Rename `UseItems.names` to `UseItems.items`.
Adjust tests and snapshots for this change.