feat(namespace): add namespaces, use global idents
Add a new namespace module.
It is responsible for building the namespace table
with valid names and their GlobalIdents.
It reports redeclarations of the same name
and undefined names that are referenced.
Similar functionality that was in the types module is removed.

The use analysis now builds a namespace.

Add common trait for Defs, Lets and StructFields.