feat(types): immutable let error
Report an error, if a let-bound name is about to be sub-typed.
Do not assert type equality for this case.

Add new Binding type and a binding map
to keep track of which names are bound by def and let.

Add a type for the namespace result and the types result.